Fix FFT/IFFT double-tracing bug
Fixes the double-tracing bug encountered with the FFT/IFFT tracing tool.
Previously, we lost the zero-frequency component (DC average pixel
brightness) such that the average intensity was zero. Taking the
absolute value, all the negative parts of the oscillating pattern
became positive, so the frequency was doubled. Now, we allow the
average value through the filter, such that we only trace either the
dark or bright fringes.

Tracing bright/dark fringes is controlled by a simple boolean variable.

Additionally, this commit adds an example interferogram to the repo for
testing/demonstration purposes.